+++ src/Components/Footer.jsx
@@ -1,10 +1,10 @@
 import React from "react";
 import FooterStyled from "../styles/js/FooterStyled";
 
-export default function Footer(props) {
+export default function Footer({text}) {
     return (
         <FooterStyled>
-                <p>{props.text}</p>
+                <p>{text}</p>
         </FooterStyled>
     )
 }
