+++ src/Components/Header.jsx
@@ -22,5 +22,5 @@ export default class Header extends React.Component {
         </div>
       </HeaderStyle>
     );
-  } // render
-} // class header
+  } 
+}
